% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{arrange_table}
\alias{arrange_table}
\title{Arrange the reshaped table}
\usage{
arrange_table(
  DT,
  row_var1,
  row_var2 = NULL,
  split_var = NULL,
  row_var1_prefix = if (!is.null(row_var2)) paste0(row_var1, ": ") else NULL,
  total_title_by_row = "Total",
  total_title_by_col = "Grand Total"
)
}
\arguments{
\item{DT}{a reshaped data.table}

\item{row_var1}{a string representing a categorical column to be represented on rows after casting the table.}

\item{row_var2}{a string representing a categorical column to be represented as sub-rows of the column indicated in the `row_var1` argument.}

\item{split_var}{a string representing a categorical column used as a reference to create more than want summary table of the same data.}

\item{row_var1_prefix}{defines the prefix to be used in summary rows related to  `row_var1` argument.}

\item{total_title_by_row}{a title to describe the column created to have the total of each row.}

\item{total_title_by_col}{a title to describe the row created to have the total of each column.}
}
\value{
A `data.table`
}
\description{
Arrange the reshaped table
}
