% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tables.R
\name{col_format}
\alias{col_format}
\title{Defines a custom format}
\usage{
col_format(numfmt, width, h_align, v_align = "center", wrap_text = FALSE)
}
\arguments{
\item{numfmt}{a string value}

\item{width}{a numeric value}

\item{h_align}{a string value}

\item{v_align}{a string value}

\item{wrap_text}{a Boolean value}
}
\value{
a named \code{list} based on column properties
}
\description{
Defines a custom format
}
