% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{custom_dcast}
\alias{custom_dcast}
\title{Long-to-wide reshaping tool}
\usage{
custom_dcast(DT, names_from, id_cols, summary_col, values_from, ...)
}
\arguments{
\item{DT}{a data.table}

\item{names_from}{string vars to be used as columns}

\item{id_cols}{string vars to be used as rows}

\item{summary_col}{rows to be used as totals}

\item{values_from}{rows to be used as values}

\item{...}{it goes to \code{dcast.data.table}}
}
\value{
A reshaped data.table
}
\description{
Long-to-wide reshaping tool
}
