% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{split_summary}
\alias{split_summary}
\title{Creates a list based on summary}
\usage{
split_summary(DT, split_var, lookup_v = NULL, name_suffix = NULL)
}
\arguments{
\item{DT}{a reshaped data.table}

\item{split_var}{a string making the reference to the column we need to use a reference to split the \code{data.table}.}

\item{lookup_v}{a named vector used to change the groups used to split the rows.}

\item{name_suffix}{adding a suffix to the table name to be show as the title of each sheet.}
}
\value{
a \code{list} of \code{data.table}
}
\description{
Divide the rows of the data.table into different elements
of a list but keeping the names provided in the \code{split_var}.
}
