% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tables.R
\name{st_add_summary}
\alias{st_add_summary}
\title{Add sheet with a summary table}
\usage{
st_add_summary(
  wb,
  DT,
  sheet_title,
  highlight_col,
  highlight_row_pattern,
  custom_col_format,
  header_custom_h_align,
  sheet_name = sheet_title,
  values_types = c("Number", "Value"),
  sheet_title_row = 1L,
  sheet_title_size = 22,
  start_row = 4L,
  start_col = 1L,
  zoom = 100,
  header_has_levels = TRUE,
  header_color_fill = openxlsx2::wb_color(hex = "FF1F497D"),
  header_color_font = openxlsx2::wb_color(name = "white"),
  header_color_border_bottom = openxlsx2::wb_color(name = "white"),
  header_bold_font = TRUE,
  header_h_align = "center",
  header_v_align = "center",
  header_wrap_text = TRUE,
  header_avoid_merge = values_types,
  num_col_pattern = "_Number$",
  num_col_numfmt = "#,##0",
  num_col_width = 10,
  num_col_h_align = "center",
  num_col_v_align = "center",
  value_col_pattern = "_Value$",
  value_col_numfmt = "$#,##0",
  value_col_width = 12,
  value_col_h_align = "center",
  value_col_v_align = "center",
  highlight_row_fill = openxlsx2::wb_color(hex = "FFD9D9D9")
)
}
\arguments{
\item{wb}{a workbook.}

\item{DT}{a reshaped \code{data.table}.}

\item{sheet_title}{a string to define the title to define the sheet.}

\item{highlight_col}{a string to define which have a pattern to highlight as summary row.}

\item{highlight_row_pattern}{the pattern to find in the \code{highlight_col} variable.}

\item{custom_col_format}{a named \code{list} based on DT var names with the results of calling the \code{col_format} function.}

\item{header_custom_h_align}{align content horizontal ('left', 'center', 'right')}

\item{sheet_name}{a string to define the sheet name.}

\item{values_types}{a}

\item{sheet_title_row}{a}

\item{sheet_title_size}{a}

\item{start_row}{a}

\item{start_col}{a}

\item{zoom}{a}

\item{header_has_levels}{a}

\item{header_color_fill}{a}

\item{header_color_font}{a}

\item{header_color_border_bottom}{a}

\item{header_bold_font}{a}

\item{header_h_align}{a}

\item{header_v_align}{a}

\item{header_wrap_text}{a}

\item{header_avoid_merge}{a}

\item{num_col_pattern}{a}

\item{num_col_numfmt}{a}

\item{num_col_width}{a}

\item{num_col_h_align}{a}

\item{num_col_v_align}{a}

\item{value_col_pattern}{a}

\item{value_col_numfmt}{a}

\item{value_col_width}{a}

\item{value_col_h_align}{a}

\item{value_col_v_align}{a}

\item{highlight_row_fill}{a}
}
\value{
a workbook object
}
\description{
This function add a summery table to an existing Excel document.
}
