% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tables.R
\name{st_add_table}
\alias{st_add_table}
\title{Add a table data in workbook}
\usage{
st_add_table(
  wb,
  DT,
  sheet_name,
  custom_col_format,
  sheet_title,
  sheet_title_row = 1L,
  sheet_title_size = 22L,
  ...
)
}
\arguments{
\item{wb}{a}

\item{DT}{a}

\item{sheet_name}{a}

\item{custom_col_format}{a}

\item{sheet_title}{a}

\item{sheet_title_row}{a}

\item{sheet_title_size}{a}

\item{...}{a}
}
\value{
a workbook object
}
\description{
Add a table data in workbook
}
