% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{summary_reshape}
\alias{summary_reshape}
\title{Creates a tidy summary}
\usage{
summary_reshape(
  DT,
  vars_to_count = NULL,
  vars_to_summary = NULL,
  row_var1,
  col_var = NULL,
  row_var2 = NULL,
  split_var = NULL,
  value_fun = function(x) sum(x, na.rm = TRUE),
  value_total_fun = value_fun,
  col_title = col_var,
  total_title_by_row = "Total",
  total_title_by_col = "Grand Total",
  col_title_sep = "_",
  reshape_summary = TRUE,
  ...
)
}
\arguments{
\item{DT}{a data.table or data.frame with the data to summarize.}

\item{vars_to_count}{a named string vector with the columns we want to apply an unique count. The names will be used to identify the values after casting.}

\item{vars_to_summary}{a named string vector with the columns we want to summary using the function defined in the \code{value_fun} argument. The names will be used to identify the values after casting.}

\item{row_var1}{a string representing a categorical column to be represented on rows after casting the table.}

\item{col_var}{a string representing a categorical column to be represented on columns after casting the table.}

\item{row_var2}{a string representing a categorical column to be represented as sub-rows of the column indicated in the \code{row_var1} argument.}

\item{split_var}{a string representing a categorical column used as a reference to create more than want summary table of the same data.}

\item{value_fun}{defines a simple summary function for the variables defined in the \code{vars_to_summary} argument. By default it calculates the sum of variables.}

\item{value_total_fun}{defines the functions to be used to calculate the Total by row and column.}

\item{col_title}{a title to describe the variable used to create the columns.}

\item{total_title_by_row}{a title to describe the column created to have the total of each row.}

\item{total_title_by_col}{a title to describe the row created to have the total of each column.}

\item{col_title_sep}{a separator needed to define hierarchy between \code{col_title} and \code{col_var}.}

\item{reshape_summary}{if \code{TRUE} reshape the data according prior arguments.}

\item{...}{it goes to \code{dcast.data.table}}
}
\value{
A \code{data.table}
}
\description{
Create all the values to be displayed in the summary table in tidy format.

If after creating the summary you need to add more columns to the final table,
it's recommended to use this tidy format before casting the final table.
}
